python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pendulum>=3.1.0",
    "pydantic>=2.10.6",
    "pydantic-extra-types>=2.10.4",
    "pyharp",
    "pyserial>=3.5",
    "pyyaml>=6.0.2",
    "speaker-calibration",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sgen", "src/shutdown"]

[tool.uv.sources]
pyharp = { git = "https://github.com/harp-tech/pyharp.git" }
speaker-calibration = { git = "ssh://git@github.com/fchampalimaud/cdc-speaker-calibration", branch = "feature/chirp" }

[project.scripts]
generate-schemas = "sgen.__main__:main"
config = "config.__main__:main"
startup = "startup.__main__:main"
convert-output = "shutdown.output_to_csv:convert_output"
merge-output = "shutdown.output_to_csv:merge_output"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
